Rename component -> di
