+++ app/di/service.go
@@ -1,4 +1,4 @@
-package component
+package di
 
 import "github.com/izumin5210/dform/app/service"
